refactoring: exportQcmExport et creerDocumentAMC sortent de outils.js
fix
